Move GracefulTerminationException to a new package
