+++ src/python/pants/base/exceptions.py
@@ -62,28 +62,3 @@ class BackendConfigurationError(BuildConfigurationError):
 
 class IncompatiblePlatformsError(Exception):
   """Indicates that target platforms are incompatible with a target that contains native code."""
-
-
-class GracefulTerminationException(Exception):
-  """Indicates that a console_rule should eagerly terminate the run.
-
-  No error trace will be printed if this is raised; the runner will simply exit with the passed
-  exit code.
-
-  Nothing except a console_rule should ever raise this.
-  """
-
-  def __init__(self, message='', exit_code=1):
-    """
-    :param int exit_code: an optional exit code (default=1)
-    """
-    super(GracefulTerminationException, self).__init__(message)
-
-    if exit_code == 0:
-      raise ValueError("Cannot create GracefulTerminationException with exit code of 0")
-
-    self._exit_code = exit_code
-
-  @property
-  def exit_code(self):
-    return self._exit_code
